move appointments queries to seperated directory
